reorganizing tests - nodejs OK - cannot get browser tests going
